+++ src/wai/common/meta/constants.py
@@ -1,0 +1,5 @@
+"""
+Module for constant values to do with Python programming.
+"""
+TRIPLE_QUOTES: str = '"""'
+"""Triple-quotes used to surround multiline strings/doc-strings."""