+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 - Added function to get the fully-qualified name of a class.
 - Added function to get the import code for a class.
 - Added ClassRegistry class.
+- Added meta-constants module.
 
 0.0.24 (2020-02-21)
 -------------------